interview/src/test/java/facebook/practice/Matching_Pairs/SolutionITest.java
package facebook.practice.Matching_Pairs;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 5/1/2021
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final String s_1 = "abcde", t_1 = "adcbe";
    int expected_1 = 5;
    int output_1 = solution.matchingPairs(s_1, t_1);
    assertEquals(expected_1, output_1);
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    String s_2 = "abcd", t_2 = "abcd";
    int expected_2 = 2;
    int output_2 = solution.matchingPairs(s_2, t_2);
    assertEquals(expected_2, output_2);
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}